+++ src/main/java/blusunrize/immersiveengineering/common/blocks/BlockIEFluid.java
@@ -14,7 +14,6 @@ import net.minecraft.block.material.Material;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityLivingBase;
-import net.minecraft.item.ItemBlock;
 import net.minecraft.potion.PotionEffect;
 import net.minecraft.util.EnumFacing;
 import net.minecraft.util.math.BlockPos;
@@ -37,9 +36,7 @@ public class BlockIEFluid extends BlockFluidClassic
 		super(fluid, material);
 		this.setUnlocalizedName(ImmersiveEngineering.MODID+"."+name);
 		this.setCreativeTab(ImmersiveEngineering.creativeTab);
-//		ImmersiveEngineering.registerBlock(this, ItemBlock.class, name);
 		IEContent.registeredIEBlocks.add(this);
-		IEContent.registeredIEItems.add(new ItemBlock(this));
 	}
 
 	public BlockIEFluid setFlammability(int flammability, int fireSpread)