+++ src/main/java/blusunrize/immersiveengineering/common/util/compat/jei/AssemblerRecipeTransferHandler.java
@@ -11,7 +11,9 @@ package blusunrize.immersiveengineering.common.util.compat.jei;
 import blusunrize.immersiveengineering.ImmersiveEngineering;
 import blusunrize.immersiveengineering.api.Lib;
 import blusunrize.immersiveengineering.common.gui.ContainerAssembler;
-import blusunrize.immersiveengineering.common.util.network.MessageTileSync;
+import blusunrize.immersiveengineering.common.util.network.MessageSetGhostSlots;
+import it.unimi.dsi.fastutil.ints.Int2ObjectMap;
+import it.unimi.dsi.fastutil.ints.Int2ObjectOpenHashMap;
 import mezz.jei.api.gui.IGuiIngredient;
 import mezz.jei.api.gui.IGuiItemStackGroup;
 import mezz.jei.api.gui.IRecipeLayout;
@@ -19,8 +21,6 @@ import mezz.jei.api.recipe.transfer.IRecipeTransferError;
 import mezz.jei.api.recipe.transfer.IRecipeTransferHandler;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
-import net.minecraft.nbt.NBTTagCompound;
-import net.minecraft.nbt.NBTTagList;
 import net.minecraft.util.NonNullList;
 import net.minecraft.util.text.translation.I18n;
 
@@ -49,7 +49,6 @@ public class AssemblerRecipeTransferHandler implements IRecipeTransferHandler<Co
 				if(doTransfer)
 				{
 					IGuiItemStackGroup stacks = recipeLayout.getItemStacks();
-					NBTTagList tagList = new NBTTagList();
 					NonNullList<ItemStack> convertedInput = NonNullList.withSize(stacks.getGuiIngredients().size()-1, ItemStack.EMPTY);
 					int j = 0;
 					for(IGuiIngredient<ItemStack> ingr : stacks.getGuiIngredients().values())
@@ -62,19 +61,10 @@ public class AssemblerRecipeTransferHandler implements IRecipeTransferHandler<Co
 						}
 						j++;
 					}
+					Int2ObjectMap<ItemStack> changes = new Int2ObjectOpenHashMap<>();
 					for(int slot = 0; slot < Math.min(convertedInput.size(), 9); slot++)
-					{
-						container.putStackInSlot(i*10+slot, convertedInput.get(slot));
-						NBTTagCompound itemTag = new NBTTagCompound();
-						if(!convertedInput.get(slot).isEmpty())
-							convertedInput.get(slot).writeToNBT(itemTag);
-						itemTag.setInteger("slot", slot);
-						tagList.appendTag(itemTag);
-					}
-					NBTTagCompound tag = new NBTTagCompound();
-					tag.setTag("patternSync", tagList);
-					tag.setInteger("recipe", i);
-					ImmersiveEngineering.packetHandler.sendToServer(new MessageTileSync(container.tile, tag));
+						changes.put(i*10+slot, convertedInput.get(slot));
+					ImmersiveEngineering.packetHandler.sendToServer(new MessageSetGhostSlots(changes));
 				}
 				return null;
 			}