+++ changelog.md
@@ -1,4 +1,5 @@
 #####Version 0.12-86
+ - Added support for JEI ghost ingredients to the assembler and routers (Malte)
  - Re-Added razor wires hurting players trying to break them without wirecutters (Malte)
  - Re-Added coloured pipes (Malte)
  - Re-Added structural arms (Malte)